strcasecmp availability
doc/examples/largepost.c uses strcasecmp which is not part of ISO C.
This will create a warning if -std=c99 is passed to gcc.

The function strcasecmp can be defined in string.h or strings.h
To check the availability of strcasecmp, we can use the autoconf macros
AX_STRING_STRCASECMP and AX_STRINGS_STRCASECMP.

As previously, under Windows, we can use _stricmp.

In all other cases, we can define our own strcasecmp as a fallback.