Add macro MHD_UNUSED
When using GCC with -Wextra (or -Wunused-parameter) a warning will be
appear when a parameter of a function is not used.
This case is very common since there are  lots of callbacks when using
this library.
The macro MHD_UNUSED removes these useless warnings. We just have to add
this macro after an unused parameter.
The MHD_UNUSED is defined as follows:
- with GCC (and a minimal version), it uses the GCC attribute 'unused'
- outisde GCC, it is the empty macro.
Therefore, MHD_UNUSED does not harm portability.

The macro MHD_UNUSED is used a little in libmicrohttpd itself, but
mostly in examples. And this removes a tons warning when using -Wextra.

Since this macro is very useful when using libmicrohttpd, it is in the
public header.

This way of removing this family of warnings is actually quite common.
For example, with SDL, there is the macro SDL_UNUSED.